add visibility check before resuming top activity from user switch
Modify the codepath for resuming top activities after a user switch so the activity that was left
open when the user is locked isn't unconditionally resumed after switching away and switching back
to the user. The keyguard visibility is checked so that apps that explicitly opt into showing an
activity when locked are still shown when switching back.

Test: atest FrameworksServicesTests:UserControllerTest WmTests:ActivityTaskManagerServiceTests WmTests:RootWindowContainerTests
Test: manual